ensembldl: incorporated latest changes from ncbidl

+++ ensembldl.sh
@@ -25,7 +25,7 @@ $app $version
 
 USAGE
 
-  $app [options] [--] dataset output
+  $app [options] [--] prefix dataset...
 
 DESCRIPTION
 
@@ -33,14 +33,14 @@ DESCRIPTION
 
 ARGUMENTS
 
-  dataset               the remote data set to download from the ftp server,
-                        example: pub/release-93/fasta/gallus_gallus/dna/Gallus_gallus
-
-  output                the local data set directory,
+  prefix                the local data set directory,
                         example: /data/db
 
-                        the data set will be put in:
-                        \$output/ensembl/\$dataset/\$(date +%F)
+                        the data sets will be put in:
+                        \$prefix/ensembl/\$dataset/\$(date +%F)
+
+  dataset...            the remote data sets to download from the ftp server,
+                        example: pub/release-93/fasta/gallus_gallus/dna/Gallus_gallus
 
   --                    ends option parsing
 
@@ -120,32 +120,40 @@ do
 done
 
 set +o nounset
-dataset=$1
-shift || bailout "missing argument: data set"
-output_prefix=$1
-shift || bailout "missing argument: output"
+prefix=$1
+shift || bailout "missing argument: prefix"
 set -o nounset
 
-if [[ "$*" != "" ]]
-then
-  bailout "trailing arguments: $*"
-fi
-
-# trim trailing slashes
 shopt -s extglob
-dataset="${dataset%%+(/)}"
-output_prefix="${output_prefix%%+(/)}"
+# trim trailing slashes
+prefix="${prefix%%+(/)}"
+
+datasets=()
+for d in "$@"
+do
+  datasets+=("${d%%+(/)}")
+done
+
 shopt -u extglob
+
+[[ ${#datasets[@]} -gt 0 ]] ||
+  bailout 'missing argument: dataset'
 
 # -----------------------------------------------------------------------------
 # verbose: output configuration
 # -----------------------------------------------------------------------------
 
-[[ $verbose == yes ]] &&
-cat << EOF
-dataset: $dataset
-output: $output_prefix
-
+if [[ $verbose == yes ]]
+then
+  cat << EOF
+prefix: $prefix
+datasets:
+EOF
+  for d in "${datasets[@]}"
+  do
+    echo "- $d"
+  done
+  cat << EOF
 parallel: $cores CPU cores
 
 versions:
@@ -153,32 +161,23 @@ versions:
 - $(lftp --version | head -1)
 
 EOF
+fi
 
 # -----------------------------------------------------------------------------
 # check arguments
 # -----------------------------------------------------------------------------
 
-[[ -n $dataset ]] ||
-  bailout "no remote data set specified"
+[[ -d $prefix ]] ||
+  bailout "local data set directory does not exist: $prefix"
 
-[[ -n $output_prefix ]] ||
-  bailout "local data set directory not specified"
-
-[[ -d $output_prefix ]] ||
-  bailout "local data set directory does not exist: $output_prefix"
+download_date=$(date +%F)
 
 # -----------------------------------------------------------------------------
 # preparation
 # -----------------------------------------------------------------------------
 
-tmpdir=$(mktemp -d --tmpdir="$output_prefix" ".$app-XXXXXXXXXX")
+tmpdir=$(mktemp -d --tmpdir="$prefix" ".$app-XXXXXXXXXX")
 trap 'rm -fr "$tmpdir"' EXIT INT TERM
-
-download_date=$(date +%F)
-
-output_basedir="$output_prefix/ensembl/$dataset"
-
-output_dir="$output_basedir/$download_date"
 
 # -----------------------------------------------------------------------------
 # application functions
@@ -212,43 +211,53 @@ EOF
 # application
 # -----------------------------------------------------------------------------
 
-[[ $verbose == yes ]] &&
-  log.info "starting download of $dataset"
+for dataset in "${datasets[@]}"
+do
+  output_basedir="$prefix/ncbi/$dataset"
 
-download ||
-  bailout 'download failed'
+  output_dir="$output_basedir/$download_date"
 
-[[ $verbose == yes ]] &&
-  log.info "checking md5 checksums"
+  if [[ -e $output_dir ]]
+  then
+    log.info "skipping $dataset: already exists"
+    continue
+  fi
 
-pushd "$tmpdir" &> /dev/null
+  [[ $verbose == yes ]] &&
+    log.info "starting download of $dataset"
+  download ||
+    bailout 'download failed'
 
-#ensembl sometimes has 3 cases for checksums: 
-# - MD5SUM file (that need to be checked with md5sum) 
-# - HKSUMS file (to be checked with sum) 
-# - no checksum at all
-if [[ -f MD5SUM ]]; then
-  find . -name "*gz" | while read -r file
-  do
-    grep $(basename $file .gz) MD5SUM
-  done | md5sum -c --quiet || bailout 'verification error'
-  rm MD5SUM
-elif [[ -f CHECKSUMS ]]; then
-  find . -name "*gz" | while read -r file
-  do
-    grep $(basename $file .gz) CHECKSUMS
-  done | sort -k 3 > _CHKSUMSAVAIL
-  (sum $(find . -iname "*gz" | sed 's/\.\///' ) | sort -k 3 | diff -q - _CHKSUMSAVAIL) ||
-  bailout 'verification error'
-  rm CHECKSUMS _CHKSUMSAVAIL
-else
-  log.info "checksums unavailable -> skipping verification"
-fi
+  [[ $verbose == yes ]] &&
+    log.info "checking md5 checksums"
 
-[[ $verbose == yes ]] &&
-  log.info "extracting files"
+  pushd "$tmpdir" &> /dev/null
 
-find . -type f |
+  #ensembl sometimes has 3 cases for checksums: 
+  # - MD5SUM file (that need to be checked with md5sum) 
+  # - HKSUMS file (to be checked with sum) 
+  # - no checksum at all
+  if [[ -f MD5SUM ]]; then
+    find . -name "*gz" | while read -r file
+    do
+      grep $(basename $file .gz) MD5SUM
+    done | md5sum -c --quiet || bailout 'verification error'
+    rm MD5SUM
+  elif [[ -f CHECKSUMS ]]; then
+    find . -name "*gz" | while read -r file
+    do
+      grep $(basename $file .gz) CHECKSUMS
+    done | sort -k 3 > _CHKSUMSAVAIL
+    (sum $(find . -iname "*gz" | sed 's/\.\///' ) | sort -k 3 | diff -q - _CHKSUMSAVAIL) ||
+    bailout 'verification error'
+    rm CHECKSUMS _CHKSUMSAVAIL
+  else
+    log.info "checksums unavailable -> skipping verification"
+  fi
+
+  [[ $verbose == yes ]] &&
+    log.info "extracting files"
+
   while read -r file
   do
     case "$file" in
@@ -264,24 +273,28 @@ find . -type f |
         ;;
 
       *)
-        bailout "do not recognize file type, open issue https://github.com/idiv-biodiversity/scddl/issues"
+        bailout << EOF
+do not recognize file type, please open issue for support:
+  https://github.com/idiv-biodiversity/scddl/issues
+EOF
         ;;
     esac
-  done
+  done < <(find . -type f)
 
-popd &> /dev/null
+  popd &> /dev/null
 
-[[ $verbose == yes ]] &&
-  log.info "moving from tmp dir to final destination"
+  [[ $verbose == yes ]] &&
+    log.info "moving from tmp dir to final destination"
 
-mkdir -p "$(dirname "$output_dir")"
+  mkdir -p "$(dirname "$output_dir")"
 
-mv -n "$tmpdir" "$output_dir" ||
-  bailout "moving failed"
+  mv -n "$tmpdir" "$output_dir" ||
+    bailout "moving failed"
 
-[[ $verbose == yes ]] &&
-  log.info "setting read only"
+  chmod -R +r "$output_dir"
+done
 
-chmod -R -w "$output_dir"
-
-log.info "done"
+if [[ $verbose == yes ]]
+then
+  log.info "done"
+fi
